admin/DisCode/requirements.txt
appnope==0.1.2
asgiref==3.4.1
attrs==21.4.0
Automat==20.2.0
autopep8==1.6.0
backcall==0.2.0
black==21.12b0
bson==0.5.8
certifi==2021.10.8
click==8.0.3
constantly==15.1.0
cypari==2.4.1
dataclasses==0.6
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.0.5
django-ckeditor==6.2.0
django-cors-headers==3.10.1
django-import-export==2.7.1
django-js-asset==1.2.2
djangorestframework==3.12.4
djongo==1.3.3
dnspython==2.1.0
et-xmlfile==1.1.0
FXrays==1.3.5
gssapi==1.7.2
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.3
import-export==0.2.67.dev6
incremental==21.3.0
ipython==7.31.0
jedi==0.18.1
Jinja2==3.0.3
knot-floer-homology==1.1
MarkupPy==1.14
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
networkx==2.6.3
odfpy==1.4.1
openpyxl==3.0.9
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
platformdirs==2.4.1
plink==2.4.1
prompt-toolkit==3.0.24
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pycodestyle==2.8.0
Pygments==2.11.2
pymongo==3.12.1
pypng==0.0.21
python-dateutil==2.8.2
python-decouple==3.6
pytz==2021.3
PyYAML==6.0
six==1.16.0
snappy==3.0.3
snappy-manifolds==1.1.2
spherogram==2.1
sqlparse==0.2.4
srv==1.1.0
tablib==3.1.0
toml==0.10.2
tomli==1.2.3
traitlets==5.1.1
Twisted==21.7.0
typing_extensions==4.0.1
wcwidth==0.2.5
whitenoise==6.1.0
xlrd==2.0.1
xlwt==1.3.0
yapf==0.32.0
zope.interface==5.4.0
